requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-pptx==0.6.21
mammoth==1.6.0
PyPDF2==3.0.1
openai==0.28.1
python-dotenv==1.0.0
reportlab==4.0.4
Pillow==10.0.1
jinja2==3.1.2
gunicorn==21.2.0
psutil==5.9.5
redis==4.6.0
celery==5.3.1
gunicorn==21.2.0 